# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": " Este projeto é um sistema de CRUD (Create, Read, Update, Delete) para cadastro de usuários, desenvolvido com Laravel, Vue.js e Docker. O sistema permite a criação, visualização, atualização e exclusão de registros de usuários em uma aplicação web moderna.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuario": {
            "get": {
                "summary": "Lista todos os usuários",
                "operationId": "5f08371670ce270a0e6577fae9539512",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Cria um novo usuário",
                "operationId": "d5e32171c128f56efd1872d1ec23f2c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usuario/{id}": {
            "get": {
                "summary": "Mostra um usuário específico",
                "operationId": "3578cec20dc77ec76383f2a46f907876",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "summary": "Atualiza um usuário existente",
                "operationId": "ea3d8c0180893fb3e107fae2b843db85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleta um usuário",
                "operationId": "ab48b3f880701f25e88fd267382fe63c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário deletado com sucesso"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer"
                    },
                    "nome": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}